Fix cluster templates update
This change fixes the following issues:

* Remove _appy_defaults for cluster template fields on update
  Fields that are not specified should not be given default values
  on update like they are on creation.  If a field is not specified,
  it should be left alone

* Do not delete node groups if the cluster template update does
  specify a node_group field

Closes-Bug: #1438077
(cherry picked from commit 196f72f18e7c5649580a7746b87e35defaadb943)

+++ sahara/db/sqlalchemy/api.py
@@ -542,13 +542,19 @@ def cluster_template_update(context, values, ignore_default=False):
             )
         cluster_template.update(values)
 
-        model_query(m.TemplatesRelation, context).filter_by(
-            cluster_template_id=cluster_template_id).delete()
-        for ng in node_groups:
-            node_group = m.TemplatesRelation()
-            node_group.update(ng)
-            node_group.update({"cluster_template_id": cluster_template_id})
-            node_group.save(session=session)
+        # If node_groups has not been specified, then we are
+        # keeping the old ones so don't delete!
+        if node_groups:
+            model_query(m.TemplatesRelation,
+                        context, session=session).filter_by(
+                cluster_template_id=cluster_template_id).delete()
+
+            for ng in node_groups:
+                node_group = m.TemplatesRelation()
+                node_group.update(ng)
+                node_group.update({"cluster_template_id":
+                                   cluster_template_id})
+                session.add(node_group)
 
     return cluster_template
